OpenSubtitle: Move Subtitle in its own file
