+++ Mpdn.Extensions.csproj
@@ -123,6 +123,8 @@
     <Compile Include="Extensions\PlayerExtensions\Interfaces\ILavAudioSettings.cs" />
     <Compile Include="Extensions\PlayerExtensions\AudioSettings.cs" />
     <Compile Include="Extensions\PlayerExtensions\Interfaces\ISubtitlesCom.cs" />
+    <Compile Include="Extensions\PlayerExtensions\OpenSubtitles\Subtitle.cs" />
+    <Compile Include="Extensions\PlayerExtensions\OpenSubtitles\SubtitleManager.cs" />
     <Compile Include="Extensions\PlayerExtensions\TaskbarEnhancer.cs" />
     <Compile Include="Extensions\PlayerExtensions\KeyBindings.cs" />
     <Compile Include="Extensions\PlayerExtensions\MouseControl.ConfigDialog.cs">
